+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # LoginLdap Changelog
 
+#### LoginLdap 5.2.4 - 2026-08-10
+- Added code to change the logic for random password generation
+
 #### LoginLdap 5.2.3 - 2026-08-03
 * Enabled password confirmation by default on the LoginLdap settings page
 * Added code to synchronize LDAP users using the resolved Matomo login instead of the supplied identifier